update response to return long ids

+++ recommendation-model/src/main/java/ua/tunepoint/recommendation/model/response/payload/RecommendationPayload.java
@@ -14,5 +14,5 @@ import java.util.List;
 public class RecommendationPayload {
 
     private User user;
-    private List<Integer> items;
+    private List<Long> items;
 }
